Update store-ios.js
In ios-adapter.js on the store.when("re-refreshed") event refreshReceipts is called with a callback function that takes one argument. This argument is considered to have an attribute of bundleIdentifier. However it is impossible for this argument to have a bundleIdentifier attribute, since at can only be the base64 variable (raw data string). So there must be a bug either in store-ios.js or in ios-adapter.js.