Fixes 398 __NSArrayI JSONRepresentation]: unrecognized selector sent to instance
In Cordova iOS 4.0, which was recently release, a fix was made to avoid
a namespace clashing recommendation from Apple.  Issue is here:
https://issues.apache.org/jira/browse/CB-8270
The fix was to rename JSONString->cdv_JSONString and
JSONRepresentation->cdv_JSONRepresentation.

WARNING: This would likely break someone running an older Cordova
version, so not sure how you want to manage this