(ios) Cache appStoreReceipt for 7 days
Having an up-to-date receipt is a good thing to check for events like
cancelled transactions and introductory-price eligibility.

However, each time we refresh the receipt, the app will ask the user for
its iTunes password (if he's not logged in with iTunes).

To mitigate that:

 * we will refresh the purchase when a purchase is made or restoring
 transactions was requested. (user already had to enter the password).
 * we give the oportunity to the receipt validator to refresh receipts
 using the _lastReceipt field. (up-to-date receipts can be requested
 server-side without asking the user for a password).

There'll also be a request for an up-to-date request at the first
application startup, which shouldn't be a problem: the user just
downloaded the app from iTunes, so he's supposed to be logged-in
already.

Lastly, we now store the app itself as a virtual product, so when a
receipt validator is in place it'll get an opportunity to update the
client's appStoreReceipt when the app starts. This way, the application
should never have to do a client-side refresh (thus never ask for a
password).