Ensure approved can't be called twice
There could be a case were a given "approved" handler is responsible for a
receipt update that would in turn trigger "approved". This fixes the issue by
storing the "date of last call to approved handler" before making the calls.
This date is used so handlers are not called more than once every minute for a
given transaction.