Fix #1434 - validProducts.find is not a product
Issue description:

It appears that on some devices, sometimes the call for getAvailableProducts
returns the string "OK" instead of the expected array, thus we get the error
validProducts.find is not a function. This is a fatal error. I think the
problem is that the mCallbackContext is stored on the PurchasePlugin class,
thus, if after calling the getAvailableProducts method, we get the
callSuccess() of a previous request which returns the status "OK", it will
answer the getAvailableProducts with the status string "OK" which leads to this
exception. I think that at least we should overcome this problem by checking
the validProducts in the response on the getAvailableProducts before using it.
If it isn't an array, just retry the getAvailableProducts function instead of
using it.

Fix:

This hopefully fixes the issue at its root by ensuring in the native side that
the callback used to respond is indeed the right one. The fix also provides the
suggested work-around, in case something unexpected happens, and the array of
products isn't an array, it won't hurt.